mlops_pipeline/requirements.txt
# AgriKD MLOps Pipeline - Python 3.10
# =============================================
# Install: pip install -r requirements.txt
#
# torch & torchvision cài riêng (cần chỉ định index URL):
#   CPU:  pip install torch torchvision --index-url https://download.pytorch.org/whl/cpu
#   CUDA: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121

# ONNX ecosystem
onnx>=1.14
onnxruntime>=1.16
onnxsim>=0.4

# TFLite conversion
ai-edge-torch>=0.2
onnx2tf>=1.20
tensorflow>=2.15

# Evaluation & Benchmarking
psutil>=5.9
pandas>=2.0
matplotlib>=3.7
seaborn>=0.12
tabulate>=0.9
scipy>=1.10
thop>=0.1
scikit-learn>=1.0

# Core
numpy>=1.21
